refactor(settings): reorganize file structure
