+++ src/components/Settings/About/About.vue
