+++ src/components/Settings/Compatibility/Compatibility.vue
@@ -2,8 +2,8 @@
 import { settings } from '~/logic'
 import { isHomePage } from '~/utils/main'
 
-import SettingsItem from './SettingsItem.vue'
-import SettingsItemGroup from './SettingsItemGroup.vue'
+import SettingsItem from '../components/SettingsItem.vue'
+import SettingsItemGroup from '../components/SettingsItemGroup.vue'
 
 watch(() => settings.value.useOriginalBilibiliHomepage, () => {
   if (isHomePage())
